pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aems-agent"
version = "0.2.0"
description = "AEMS Local Bridge Agent — local filesystem access for exam PDFs"
authors = [{name = "AEMS Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
tray = ["pystray>=0.19.0", "pillow>=10.0.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "httpx>=0.25.0",
]

[project.scripts]
aems-agent = "aems_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aems_agent"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "fastapi",
    "fastapi.*",
    "uvicorn",
    "uvicorn.*",
    "httpx",
    "pystray",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
